+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Incompatible changes:
 
 - CHANGED: Domain.ExpiresOn has been replaced by Domain.ExpiresAt. (dnsimple/dnsimple-go#98)
+- CHANGED: Certificate.ExpiresOn has been replaced by Certificate.ExpiresAt. (dnsimple/dnsimple-go#99)
 
 #### Release 0.63.0